Docker/exercise-01/requirements.txt
numpy == 1.25.2
Pillow == 10.0.0